renamed outputFormat.ts to exchangeData.ts
